Split VariablesHelpers.hpp slicing into multiple files
