+++ src/DataStructures/DataBox/DataOnSlice.hpp
@@ -10,8 +10,8 @@
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/Index.hpp"
 #include "DataStructures/SliceIterator.hpp"
+#include "DataStructures/SliceTensorToVariables.hpp"
 #include "DataStructures/Variables.hpp"
-#include "DataStructures/VariablesHelpers.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
