+++ src/NumericalAlgorithms/DiscontinuousGalerkin/Actions/ApplyFluxes.hpp
@@ -9,7 +9,7 @@
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/DataBox/DataBoxTag.hpp"
 #include "DataStructures/DataBox/Prefixes.hpp"
-#include "DataStructures/VariablesHelpers.hpp"
+#include "DataStructures/SliceVariables.hpp"
 #include "Domain/IndexToSliceAt.hpp"
 #include "NumericalAlgorithms/DiscontinuousGalerkin/FluxCommunicationTypes.hpp"
 #include "NumericalAlgorithms/DiscontinuousGalerkin/MortarHelpers.hpp"
