+++ tests/Unit/DataStructures/Test_Variables.cpp
@@ -18,7 +18,6 @@
 #include "DataStructures/ModalVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Variables.hpp"
-#include "DataStructures/VariablesHelpers.hpp"
 #include "ErrorHandling/Error.hpp"  // IWYU pragma: keep
 #include "Utilities/GetOutput.hpp"
 #include "Utilities/Gsl.hpp"
@@ -28,6 +27,7 @@
 #include "Utilities/TaggedTuple.hpp"
 #include "Utilities/TypeTraits.hpp"
 #include "tests/Unit/DataStructures/DataBox/TestHelpers.hpp"
+#include "tests/Unit/DataStructures/TestTags.hpp"
 #include "tests/Unit/TestHelpers.hpp"
 #include "tests/Utilities/MakeWithRandomValues.hpp"
 
@@ -36,49 +36,6 @@
 // IWYU pragma: no_include "DataStructures/VariablesForwardDecl.hpp"
 
 // IWYU pragma: no_forward_declare Variables
-
-namespace VariablesTestTags_detail {
-/// [simple_variables_tag]
-template <typename VectorType>
-struct tensor : db::SimpleTag {
-  using type = tnsr::I<VectorType, 3, Frame::Grid>;
-};
-/// [simple_variables_tag]
-template <typename VectorType>
-struct scalar : db::SimpleTag {
-  using type = Scalar<VectorType>;
-};
-template <typename VectorType>
-struct scalar2 : db::SimpleTag {
-  using type = Scalar<VectorType>;
-};
-
-/// [prefix_variables_tag]
-template <class Tag>
-struct Prefix0 : db::PrefixTag, db::SimpleTag {
-  using type = db::const_item_type<Tag>;
-  using tag = Tag;
-};
-/// [prefix_variables_tag]
-
-template <class Tag>
-struct Prefix1 : db::PrefixTag, db::SimpleTag {
-  using type = db::const_item_type<Tag>;
-  using tag = Tag;
-};
-
-template <class Tag>
-struct Prefix2 : db::PrefixTag, db::SimpleTag {
-  using type = db::const_item_type<Tag>;
-  using tag = Tag;
-};
-
-template <class Tag>
-struct Prefix3 : db::PrefixTag, db::SimpleTag {
-  using type = db::const_item_type<Tag>;
-  using tag = Tag;
-};
-}  // namespace VariablesTestTags_detail
 
 static_assert(
     std::is_nothrow_move_constructible<Variables<
@@ -810,172 +767,6 @@ void test_variables_extract_subset() noexcept {
 }
 
 template <typename VectorType>
-void test_variables_slice() noexcept {
-  MAKE_GENERATOR(gen);
-  UniformCustomDistribution<size_t> sdist{5, 10};
-
-  const size_t x_extents = sdist(gen);
-  const size_t y_extents = sdist(gen);
-  const size_t z_extents = sdist(gen);
-  Variables<tmpl::list<VariablesTestTags_detail::tensor<VectorType>>> vars{
-      x_extents * y_extents * z_extents};
-  const size_t tensor_size =
-      VariablesTestTags_detail::tensor<VectorType>::type::size();
-  Index<3> extents(x_extents, y_extents, z_extents);
-
-  // Test data_on_slice function by using a predictable data set where each
-  // entry is assigned a value equal to its index
-  for (size_t s = 0; s < vars.size(); ++s) {
-    // clang-tidy: do not use pointer arithmetic
-    vars.data()[s] = s;  // NOLINT
-  }
-  Variables<tmpl::list<VariablesTestTags_detail::tensor<VectorType>>>
-      expected_vars_sliced_in_x(y_extents * z_extents, 0.),
-      expected_vars_sliced_in_y(x_extents * z_extents, 0.),
-      expected_vars_sliced_in_z(x_extents * y_extents, 0.);
-  const size_t x_offset = sdist(gen) % x_extents;
-  const size_t y_offset = sdist(gen) % y_extents;
-  const size_t z_offset = sdist(gen) % z_extents;
-
-  for (size_t s = 0; s < expected_vars_sliced_in_x.size(); ++s) {
-    // clang-tidy: do not use pointer arithmetic
-    expected_vars_sliced_in_x.data()[s] = x_offset + s * x_extents;  // NOLINT
-  }
-  for (size_t i = 0; i < tensor_size; ++i) {
-    for (size_t x = 0; x < x_extents; ++x) {
-      for (size_t z = 0; z < z_extents; ++z) {
-        // clang-tidy: do not use pointer arithmetic
-        expected_vars_sliced_in_y
-            .data()[x + x_extents * (z + z_extents * i)] =  // NOLINT
-            i * extents.product() + x + x_extents * (y_offset + z * y_extents);
-      }
-    }
-  }
-  for (size_t i = 0; i < tensor_size; ++i) {
-    for (size_t x = 0; x < x_extents; ++x) {
-      for (size_t y = 0; y < y_extents; ++y) {
-        // clang-tidy: do not use pointer arithmetic
-        expected_vars_sliced_in_z
-            .data()[x + x_extents * (y + y_extents * i)] =  // NOLINT
-            i * extents.product() + x + x_extents * (y + y_extents * z_offset);
-      }
-    }
-  }
-
-  CHECK(data_on_slice(get<VariablesTestTags_detail::tensor<VectorType>>(vars),
-                      extents, 0, x_offset) ==
-        get<VariablesTestTags_detail::tensor<VectorType>>(
-            expected_vars_sliced_in_x));
-  CHECK(data_on_slice(get<VariablesTestTags_detail::tensor<VectorType>>(vars),
-                      extents, 1, y_offset) ==
-        get<VariablesTestTags_detail::tensor<VectorType>>(
-            expected_vars_sliced_in_y));
-  CHECK(data_on_slice(get<VariablesTestTags_detail::tensor<VectorType>>(vars),
-                      extents, 2, z_offset) ==
-        get<VariablesTestTags_detail::tensor<VectorType>>(
-            expected_vars_sliced_in_z));
-
-  CHECK(data_on_slice(vars, extents, 0, x_offset) == expected_vars_sliced_in_x);
-  CHECK(data_on_slice(vars, extents, 1, y_offset) == expected_vars_sliced_in_y);
-  CHECK(data_on_slice(vars, extents, 2, z_offset) == expected_vars_sliced_in_z);
-
-  CHECK(data_on_slice<VariablesTestTags_detail::tensor<VectorType>>(
-            extents, 0, x_offset,
-            get<VariablesTestTags_detail::tensor<VectorType>>(vars)) ==
-        expected_vars_sliced_in_x);
-  CHECK(data_on_slice<VariablesTestTags_detail::tensor<VectorType>>(
-            extents, 1, y_offset,
-            get<VariablesTestTags_detail::tensor<VectorType>>(vars)) ==
-        expected_vars_sliced_in_y);
-  CHECK(data_on_slice<VariablesTestTags_detail::tensor<VectorType>>(
-            extents, 2, z_offset,
-            get<VariablesTestTags_detail::tensor<VectorType>>(vars)) ==
-        expected_vars_sliced_in_z);
-}
-
-template <typename VectorType>
-void test_variables_add_slice_to_data() noexcept {
-  MAKE_GENERATOR(gen);
-  UniformCustomDistribution<
-      tt::get_fundamental_type_t<typename VectorType::value_type>>
-      dist{-100.0, 100.0};
-
-  // Test adding two slices on different 'axes' to a Variables
-  std::array<VectorType, 3> orig_vals;
-  std::fill(orig_vals.begin(), orig_vals.end(), VectorType{8});
-  fill_with_random_values(make_not_null(&orig_vals), make_not_null(&gen),
-                          make_not_null(&dist));
-
-  std::array<VectorType, 3> slice0_vals;
-  std::fill(slice0_vals.begin(), slice0_vals.end(), VectorType{4});
-  fill_with_random_values(make_not_null(&slice0_vals), make_not_null(&gen),
-                          make_not_null(&dist));
-
-  std::array<VectorType, 3> slice1_vals;
-  std::fill(slice1_vals.begin(), slice1_vals.end(), VectorType{2});
-  fill_with_random_values(make_not_null(&slice1_vals), make_not_null(&gen),
-                          make_not_null(&dist));
-
-  using Tensor = typename VariablesTestTags_detail::tensor<VectorType>::type;
-  const Index<2> extents{{{4, 2}}};
-  Variables<tmpl::list<VariablesTestTags_detail::tensor<VectorType>>> vars(
-      extents.product());
-  get<VariablesTestTags_detail::tensor<VectorType>>(vars) =
-      Tensor{{{orig_vals[0], orig_vals[1], orig_vals[2]}}};
-  {
-    const auto slice_extents = extents.slice_away(0);
-    Variables<tmpl::list<VariablesTestTags_detail::tensor<VectorType>>> slice(
-        slice_extents.product(), 0.);
-    get<VariablesTestTags_detail::tensor<VectorType>>(slice) =
-        Tensor{{{slice1_vals[0], slice1_vals[1], slice1_vals[2]}}};
-    add_slice_to_data(make_not_null(&vars), slice, extents, 0, 2);
-  }
-
-  {
-    const auto slice_extents = extents.slice_away(1);
-    Variables<tmpl::list<VariablesTestTags_detail::tensor<VectorType>>> slice(
-        slice_extents.product(), 0.);
-    get<VariablesTestTags_detail::tensor<VectorType>>(slice) =
-        Tensor{{{slice0_vals[0], slice0_vals[1], slice0_vals[2]}}};
-    add_slice_to_data(make_not_null(&vars), slice, extents, 1, 1);
-  }
-
-  // The slice0_vals should have been added to the second half of each of the
-  // three vectors in the tensor. The slice1_vals should have been added to
-  // entries 2 and 6 in each vector.
-  // clang-format off
-  const Tensor expected{
-      {{{orig_vals[0].at(0),
-         orig_vals[0].at(1),
-         orig_vals[0].at(2) + slice1_vals[0].at(0),
-         orig_vals[0].at(3),
-         orig_vals[0].at(4) + slice0_vals[0].at(0),
-         orig_vals[0].at(5) + slice0_vals[0].at(1),
-         orig_vals[0].at(6) + slice0_vals[0].at(2) + slice1_vals[0].at(1),
-         orig_vals[0].at(7) + slice0_vals[0].at(3)},
-        {orig_vals[1].at(0),
-         orig_vals[1].at(1),
-         orig_vals[1].at(2) + slice1_vals[1].at(0),
-         orig_vals[1].at(3),
-         orig_vals[1].at(4) + slice0_vals[1].at(0),
-         orig_vals[1].at(5) + slice0_vals[1].at(1),
-         orig_vals[1].at(6) + slice0_vals[1].at(2) + slice1_vals[1].at(1),
-         orig_vals[1].at(7) + slice0_vals[1].at(3)},
-        {orig_vals[2].at(0),
-         orig_vals[2].at(1),
-         orig_vals[2].at(2) + slice1_vals[2].at(0),
-         orig_vals[2].at(3),
-         orig_vals[2].at(4) + slice0_vals[2].at(0),
-         orig_vals[2].at(5) + slice0_vals[2].at(1),
-         orig_vals[2].at(6) + slice0_vals[2].at(2) + slice1_vals[2].at(1),
-         orig_vals[2].at(7) + slice0_vals[2].at(3)}}}};
-  // clang-format on
-
-  CHECK_ITERABLE_APPROX(
-      expected, get<VariablesTestTags_detail::tensor<VectorType>>(vars));
-}
-
-template <typename VectorType>
 void test_variables_from_tagged_tuple() noexcept {
   using value_type = typename VectorType::value_type;
   MAKE_GENERATOR(gen);
@@ -1055,18 +846,6 @@ SPECTRE_TEST_CASE("Unit.DataStructures.Variables", "[DataStructures][Unit]") {
     test_variables_extract_subset<DataVector>();
     test_variables_extract_subset<ModalVector>();
   }
-  SECTION("Test Variables slice utilities") {
-    test_variables_slice<ComplexDataVector>();
-    test_variables_slice<ComplexModalVector>();
-    test_variables_slice<DataVector>();
-    test_variables_slice<ModalVector>();
-  }
-  SECTION("Test adding slice values to Variables") {
-    test_variables_add_slice_to_data<ComplexDataVector>();
-    test_variables_add_slice_to_data<ComplexModalVector>();
-    test_variables_add_slice_to_data<DataVector>();
-    test_variables_add_slice_to_data<ModalVector>();
-  }
   SECTION("Test variables_from_tagged_tuple") {
     // The commented functions require a fix to issue #1420.
     // test_variables_from_tagged_tuple<ComplexDataVector>();
@@ -1105,40 +884,6 @@ SPECTRE_TEST_CASE("Unit.DataStructures.Variables", "[DataStructures][Unit]") {
   Variables<tmpl::list<VariablesTestTags_detail::scalar<DataVector>>> vars;
   get<VariablesTestTags_detail::scalar<DataVector>>(vars) =
       Scalar<DataVector>{{{{0.}}}};
-  ERROR("Failed to trigger ASSERT in an assertion test");
-#endif
-}
-
-// [[OutputRegex, volume_vars has wrong number of grid points.
-//  Expected 8, got 10]]
-[[noreturn]] SPECTRE_TEST_CASE(
-    "Unit.DataStructures.Variables.add_slice_to_data.BadSize.volume",
-    "[DataStructures][Unit]") {
-  ASSERTION_TEST();
-#ifdef SPECTRE_DEBUG
-  Variables<tmpl::list<VariablesTestTags_detail::tensor<DataVector>>> vars(10,
-                                                                           0.);
-  const Variables<tmpl::list<VariablesTestTags_detail::tensor<DataVector>>>
-      slice(2, 0.);
-  add_slice_to_data(make_not_null(&vars), slice, Index<2>{{{4, 2}}}, 0, 0);
-  ERROR("Failed to trigger ASSERT in an assertion test");
-#endif
-}
-
-    // clang-format off
-// [[OutputRegex, vars_on_slice has wrong number of grid points.
-//  Expected 2, got 5]]
-[[noreturn]] SPECTRE_TEST_CASE(
-    "Unit.DataStructures.Variables.add_slice_to_data.BadSize.slice",
-    "[DataStructures][Unit]") {
-  // clang-format on
-  ASSERTION_TEST();
-#ifdef SPECTRE_DEBUG
-  Variables<tmpl::list<VariablesTestTags_detail::tensor<DataVector>>> vars(8,
-                                                                           0.);
-  const Variables<tmpl::list<VariablesTestTags_detail::tensor<DataVector>>>
-      slice(5, 0.);
-  add_slice_to_data(make_not_null(&vars), slice, Index<2>{{{4, 2}}}, 0, 0);
   ERROR("Failed to trigger ASSERT in an assertion test");
 #endif
 }
