+++ tests/Unit/NumericalAlgorithms/LinearOperators/Test_Filtering.cpp
@@ -14,7 +14,6 @@
 #include "DataStructures/Matrix.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Variables.hpp"
-#include "DataStructures/VariablesHelpers.hpp"
 #include "Domain/Mesh.hpp"
 #include "Domain/Tags.hpp"  // IWYU pragma: keep
 #include "NumericalAlgorithms/LinearOperators/ApplyMatrices.hpp"
